util/slice/slice_func.go
package slice

// Union 两个slice求并集
func Union(slice1, slice2 []string) []string {
	m := make(map[string]int)
	for _, v := range slice1 {
		m[v]++
	}
	for _, v := range slice2 {
		times, _ := m[v]
		if times == 0 {
			slice1 = append(slice1, v)
		}
	}
	return slice1
}

// Intersect 两个slice求交集
func Intersect(slice1, slice2 []string) []string {
	m := make(map[string]int)
	nn := make([]string, 0)
	for _, v := range slice1 {
		m[v]++
	}
	for _, v := range slice2 {
		times, _ := m[v]
		if times == 1 {
			nn = append(nn, v)
		}
	}
	return nn
}

// Difference 两个slice求差集
// 描述：返回属于slice1，不属于slice2的差集，详情：slice_fun_test.go测试用例
// 如果想要属于slice2的差集，将slice1和slice2参数交换位置
// 如果想要两个集合的差集的并集，使用两次Difference处理，交换集合的顺序，将两次结果合并即可
// 这样做的原因：可以灵活自由的组合此处的3个
func Difference(slice1, slice2 []string) []string {
	m := make(map[string]int)
	nn := make([]string, 0)
	inter := Intersect(slice1, slice2)
	for _, v := range inter {
		m[v]++
	}
	for _, value := range slice1 {
		times, _ := m[value]
		if times == 0 {
			nn = append(nn, value)
		}
	}
	return nn
}
